Largely replaces Angular DI dependency with Constructable
A single superclass "Constructable" provides a template factory method
called "new" - factory functions based on calling .new on the target
class mean that only the top level .new needs to be exposed.

All classes are mapped by a single ES6 module, which has to be made
available to .new because Traceur's module system doesn't allow for
cyclic imports (the classmap has to refer to every other class, which
all extend (and therefore import) the Constructor class. In a complete
ES6 module system (e.g. systemjs), Constructor could import the
classmap directly, although it should still be assigned to instances so
that it could be overridden for testing purposes.

This remains an intermediate step - the artifacts of Angular DI remain.
In a couple of places, a nicer approach needed to be taken: overriding
the static new in order to provide an analogue to angular.module.value -
c.f. ListRelationshipInitializer.

With a little extra effort, individual classes can be shifted to this
approach, or to using their classMap's directly, like the
TopLevelOrchestrator does. The construct method might even be replaced
by `classMap.Something.new(args)` if the classMap didn't need to be
injected. However, existing tests rely on the factory functions
themselves being passed in as an argument, and I didn't want to spend
the time up front to convert them over.

A further thought is that if a class were converted to do local
factorying from explicit classes, then the set of classes passed into
its constructor could serve as a means of composition, and the class
heirarchy could be greatly simplified.